+++ backend/migrations/20240112131322_git_sync_multi.up.sql
@@ -1,0 +1,2 @@
+-- Add up migration script here
+UPDATE workspace_settings SET git_sync = '[]'::jsonb || git_sync