+++ frontend/src/lib/components/apps/editor/component/Component.svelte
@@ -68,6 +68,7 @@
 	import AppDecisionTree from '../../components/layout/AppDecisionTree.svelte'
 	import AppAgCharts from '../../components/display/charts/AppAgCharts.svelte'
 	import AppDbExplorer from '../../components/display/dbtable/AppDbExplorer.svelte'
+	import AppS3FileInput from '../../components/inputs/AppS3FileInput.svelte'
 
 	export let component: AppComponent
 	export let selected: boolean
@@ -613,6 +614,13 @@
 			/>
 		{:else if component.type === 'fileinputcomponent'}
 			<AppFileInput
+				configuration={component.configuration}
+				id={component.id}
+				customCss={component.customCss}
+				{render}
+			/>
+		{:else if component.type === 's3fileinputcomponent'}
+			<AppS3FileInput
 				configuration={component.configuration}
 				id={component.id}
 				customCss={component.customCss}