+++ frontend/src/lib/components/apps/editor/settingsPanel/inputEditor/StaticInputEditor.svelte
@@ -61,7 +61,7 @@
 			<IconSelectInput bind:componentInput />
 		{:else if fieldType === 'tab-select'}
 			<TabSelectInput bind:componentInput />
-		{:else if fieldType === 'resource'}
+		{:else if fieldType === 'resource' && subFieldType !== 's3'}
 			<ResourcePicker
 				initialValue={componentInput.value?.split('$res:')?.[1] || ''}
 				on:change={(e) => {
@@ -76,6 +76,21 @@
 				}}
 				showSchemaExplorer
 				resourceType="postgresql"
+			/>
+		{:else if fieldType === 'resource' && subFieldType === 's3'}
+			<ResourcePicker
+				initialValue={componentInput.value?.split('$res:')?.[1] || ''}
+				on:change={(e) => {
+					let path = e.detail
+					if (componentInput) {
+						if (path) {
+							componentInput.value = `$res:${path}`
+						} else {
+							componentInput.value = undefined
+						}
+					}
+				}}
+				resourceType="s3"
 			/>
 		{:else if fieldType === 'labeledresource'}
 			{#if componentInput?.value && typeof componentInput?.value == 'object' && 'label' in componentInput?.value && (componentInput.value?.['value'] == undefined || typeof componentInput.value?.['value'] == 'string')}