+++ frontend/src/routes/(root)/(logged)/runs/[...path]/+page.svelte
@@ -31,6 +31,7 @@
 	import { goto } from '$app/navigation'
 	import RunsQueue from '$lib/components/runs/RunsQueue.svelte'
 	import { forLater } from '$lib/forLater'
+	import { twMerge } from 'tailwind-merge'
 
 	let jobs: Job[] | undefined
 	let intervalId: NodeJS.Timeout | undefined
@@ -436,7 +437,14 @@
 		<div class="px-2">
 			<div class="flex items-center space-x-2 flex-row justify-between">
 				<div class="flex flex-row flex-wrap justify-between py-2 my-4 px-4 gap-1 items-center">
-					<h1 class="!text-2xl font-semibold leading-6 tracking-tight"> Runs </h1>
+					<h1
+						class={twMerge(
+							'!text-2xl font-semibold leading-6 tracking-tight',
+							$userStore?.operator ? 'pl-10' : ''
+						)}
+					>
+						Runs
+					</h1>
 
 					<Tooltip
 						documentationLink="https://www.windmill.dev/docs/core_concepts/monitor_past_and_future_runs"